+++ src/app/commands/help/FeatureRequest.ts
@@ -1,12 +1,13 @@
 import { CommandContext, SlashCommand, SlashCreator } from 'slash-create';
 import Log, { LogUtils } from '../../utils/Log';
 import { command } from '../../utils/SentryUtils';
+import constants from '../../service/constants/constants';
 
 export default class FeatureRequest extends SlashCommand {
 	constructor(creator: SlashCreator) {
 		super(creator, {
 			name: 'feature-request',
-			description: 'Pull up the form to submit a new feature request',
+			description: 'Pull up the form to submit a new feature request.',
 			throttling: {
 				usages: 2,
 				duration: 1,
@@ -20,8 +21,7 @@ export default class FeatureRequest extends SlashCommand {
 		LogUtils.logCommandStart(ctx);
 		// Ignore commands from bots
 		if (ctx.user.bot) return;
-
-		const form = 'https://discord.gg/yHTQkERpsD';
-		await ctx.send(`Easily submit requests at #request-features on our support discord: ${form}`).catch(Log.error);
+		
+		await ctx.send(`Easily submit requests at #request-features on our support discord: ${constants.FEATURE_REQUEST_CHANNEL_INVITE}`).catch(Log.error);
 	}
 }