+++ src/app/service/constants/apiKeys.ts
@@ -1,5 +1,5 @@
 const apiKeys = Object.freeze({
-	DISCORD_BOT_ID: process.env.DISCORD_BOT_APPLICATION_ID,
+	DISCORD_BOT_ID: process.env.DISCORD_BOT_APPLICATION_ID as string,
 	
 	twitterAppToken: process.env.TWITTER_API_TOKEN,
 	twitterAppSecret: process.env.TWITTER_API_SECRET,